project-02-alphavantage_etl/requirements.txt
requests
python-dotenv
pandas
python-telegram-bot
matplotlib